move articles from client to server
