+++ app/articles/page.tsx
@@ -2,7 +2,8 @@ import { getSEOProps } from "@/lib/seo";
 import { fetchTinaData } from "@/services/tina/fetchTinaData";
 import { dehydrate, DehydratedState, QueryClient } from "@tanstack/react-query";
 import { TinaClient } from "app/tina-client";
-import { ARTICLES_QUERY_KEY, getArticles } from "hooks/useFetchArticles";
+import { ARTICLES_QUERY_KEY } from "hooks/useFetchArticles";
+import { getArticles } from "@/services/server/articles";
 import { Metadata } from "next";
 import client from "tina/__generated__/client";
 import ArticlesIndexPage from "./index";
