+++ hooks/useFetchArticles.ts
@@ -1,18 +1,7 @@
-import { GetArticlesQueryQuery } from "@/tina/types";
 import { useInfiniteQuery } from "@tanstack/react-query";
-import client from "../tina/__generated__/client";
+import { getArticles } from "../services/server/articles";
+
 export const ARTICLES_QUERY_KEY = "articlesKey";
-
-export const getArticles = async ({
-  pageParam,
-}): Promise<GetArticlesQueryQuery> => {
-  const res = await client.queries.getArticlesQuery({
-    top: 10,
-    after: pageParam,
-  });
-
-  return res.data;
-};
 
 export const useFetchArticles = () => {
   const { data, fetchNextPage, isFetchingNextPage, error, isLoading } =
